refactor(cognitarium): move converters into own module
