+++ contracts/okp4-cognitarium/src/rdf/mod.rs
@@ -1,7 +1,9 @@
 mod atom;
+mod mapper;
 mod serde;
 mod uri;
 
 pub use self::atom::*;
+pub use self::mapper::*;
 pub use self::serde::*;
 pub use self::uri::*;
